% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerdiagnostic.R
\name{lmerdiagnostic}
\alias{lmerdiagnostic}
\title{lmerdiagnostic}
\usage{
lmerdiagnostic(d, m, iv, dv)
}
\arguments{
\item{d}{a data frame.}

\item{m}{a lmer model i.e "tr + p + (1|a)".}

\item{iv}{indepent variable of the model to compute the emmeans i.e. "tr".}

\item{dv}{a vector with the dependent variable of interest i.e c("var1","var2","var3").}
}
\value{
A serie of tables and plots useful for model diagnostic of linear and linear mixed models.
}
\description{
lmerdiagnostic
}
