% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmertableloop.R
\name{lmertableloop}
\alias{lmertableloop}
\title{lmertableloop}
\usage{
lmertableloop(d, m, ivm, ivp, dv, p, s, con, ivc, let = TRUE)
}
\arguments{
\item{d}{a data frame.}

\item{m}{a lmer model i.e "tr + p + (1|a)" or "tr * p + (1|a)".}

\item{ivm}{indepent variable of the model to compute the p-values i.e. "tr" or c("tr","p", "tr:p")}

\item{dv}{a vector with the dependent variable of interest i.e c("var1","var2","var3")}

\item{p}{significant digits of p-values.
empty:  will round the number to 3 decimals.
"full": will leave all the digits.
"std1": will round to 2 digits, or print <0.01 or <0.001.}

\item{s}{significant digits of least square means.
empty:  will round the number to 3 decimals.
number 1 to 6: will return the desired significant digits.
"full": will leave all the digits.}

\item{con}{the list of contrast... need to specify 1 independent variable.}

\item{ivc}{independent variable of the model to computer the contrasts}

\item{let}{TRUE if you want to include the letter of the tuckey's test difference next to the least square means.}
}
\value{
A data frame with variable names, least square means (emmeans), maximim SEM, p-values, and p-values of contrasts.
}
\description{
lmertableloop
}
