% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmeans.R
\name{plotmeans}
\alias{plotmeans}
\title{plotmeans1}
\usage{
plotmeans(d, m, iv, ivp, dv, plt = TRUE)
}
\arguments{
\item{d}{a data frame.}

\item{m}{a lmer model i.e "tr + p + (1|a)".}

\item{iv}{indepent variable of the model to compute the emmeans and standard error of the mean i.e. "tr".}

\item{ivp}{in case there are multiple iv for the test, you can select a single one of interest to show significant difference of letters.
If not selected it will use the first one provided.}

\item{dv}{1 dependent variable of interest i.e "var1"}
}
\value{
A column plot with least square means (column) and the standard error (error bars) and the statistical difference with a Tukey test (letters). Letters will be black if the anova's p-value for the model is significant, if the anova return p-value between 0.05 and 0.10 letters are orange, and if the p-value is above 0.10 the letters are red.
}
\description{
plotmeans1
}
