% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raneftableloop.R
\name{raneftableloop}
\alias{raneftableloop}
\title{raneftableloop}
\usage{
raneftableloop(d, m, dv, s)
}
\arguments{
\item{d}{a data frame.}

\item{m}{a lmer model i.e "tr + p + (1|a)".}

\item{dv}{a vector with the dependent variable of interest i.e c("var1","var2","var3").}

\item{s}{significant digits of least square means.
empty:  will round the number to 3 decimals.
number 1 to 6: will return the desired significant digits.}
}
\value{
A data frame with the individual random effects from the ranef function for a serie of variables.
}
\description{
raneftableloop
}
