% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorrtableloop.R
\name{VarCorrtableloop}
\alias{VarCorrtableloop}
\title{VarCorr table loop}
\usage{
VarCorrtableloop(d, m, dv, s, r)
}
\arguments{
\item{d}{a data frame.}

\item{m}{a lmer model i.e "tr + p + (1|a)".}

\item{dv}{a vector with the dependent variable of interest i.e c("var1","var2","var3")}

\item{s}{significant digits of least square means.
empty:  will round the number to 3 decimals.
number 1 to 6: will return the desired significant digits.}

\item{r}{type of residual to display.
empty:  will return both random effect and error residuals
"residuals" or "residual" or "res": will return only error residuals.
"effect" or "eff": will return random error residuals.}
}
\value{
A data frame with the standard error of the random effect and error residuals. Use the VarCorr function to return the output for a serie of variable in one linear mixed model.
}
\description{
VarCorr table loop
}
